requirements.txt
memory-profiler==0.61.0
google-cloud-storage==2.16.0
pandas==2.2.1
numpy==1.26.4